+++ budget-binder-server/src/main/kotlin/de/hsfl/budgetBinder/server/routes/AuthRoutes.kt
@@ -6,7 +6,7 @@ import de.hsfl.budgetBinder.common.ErrorModel
 import de.hsfl.budgetBinder.common.User
 import de.hsfl.budgetBinder.server.models.UserPrincipal
 import de.hsfl.budgetBinder.server.services.JWTService
-import de.hsfl.budgetBinder.server.services.UserService
+import de.hsfl.budgetBinder.server.services.interfaces.UserService
 import io.ktor.application.*
 import io.ktor.auth.*
 import io.ktor.http.*
