+++ budget-binder-server/src/main/kotlin/de/hsfl/budgetBinder/server/routes/EntryRoutes.kt
@@ -6,7 +6,7 @@ import de.hsfl.budgetBinder.common.Entry
 import de.hsfl.budgetBinder.common.ErrorModel
 import de.hsfl.budgetBinder.server.models.UserPrincipal
 import de.hsfl.budgetBinder.server.repository.parseParameterToLocalDateTime
-import de.hsfl.budgetBinder.server.services.EntryService
+import de.hsfl.budgetBinder.server.services.interfaces.EntryService
 import io.ktor.application.*
 import io.ktor.auth.*
 import io.ktor.request.*
