+++ budget-binder-server/src/main/kotlin/de/hsfl/budgetBinder/server/services/implementations/CategoryServiceImpl.kt
@@ -1,9 +1,10 @@
-package de.hsfl.budgetBinder.server.services
+package de.hsfl.budgetBinder.server.services.implementations
 
 import de.hsfl.budgetBinder.common.Category
 import de.hsfl.budgetBinder.server.models.CategoryEntity
 import de.hsfl.budgetBinder.server.models.UserEntity
 import de.hsfl.budgetBinder.server.repository.isCreatedAndEndedCorrectPeriod
+import de.hsfl.budgetBinder.server.services.interfaces.CategoryService
 import org.jetbrains.exposed.sql.transactions.transaction
 import java.time.LocalDateTime
